refactor: moved ApplicationSuccess in shared folder
