+++ src/shared/components/molecules/ApplicationSuccess/index.tsx
