+++ src/shared/components/molecules/index.ts
@@ -1,3 +1,4 @@
+export { default as ApplicationSuccess } from './ApplicationSuccess';
 export { default as EnrollSection } from './EnrollSection';
 export { default as Footer } from './Footer';
 export { default as Navbar } from './Navbar';
